# Conflicts:
#	src/main/java/com/energyict/protocolimpl/dlms/as220/emeter/AS220ImageTransfer.java
#	src/main/java/com/energyict/protocolimpl/dlms/eictz3/EictZ3.java
#	src/main/java/com/energyict/protocolimpl/dlms/g3/messaging/G3Messaging.java
#	src/main/java/com/energyict/protocolimpl/dlms/idis/IDISMessageHandler.java
#	src/main/java/com/energyict/protocolimpl/dlms/prime/messaging/FirmwareUpgrade.java
#	src/main/java/com/energyict/protocolimpl/messaging/messages/AnnotatedFWUpdateMessageBuilder.java
#	src/main/java/com/energyict/protocolimplv2/edp/messages/EDPMessageExecutor.java
#	src/main/java/com/energyict/protocolimplv2/edp/messages/EDPMessaging.java
#	src/main/java/com/energyict/protocolimplv2/messages/convertor/AS300MessageConverter.java
#	src/main/java/com/energyict/protocolimplv2/messages/convertor/Dsmr23MessageConverter.java
#	src/main/java/com/energyict/protocolimplv2/messages/convertor/Dsmr50MessageConverter.java
#	src/main/java/com/energyict/protocolimplv2/messages/convertor/EIWebPlusMessageConverter.java
#	src/main/java/com/energyict/protocolimplv2/messages/convertor/G3MeterMessageConverter.java
#	src/main/java/com/energyict/protocolimplv2/messages/convertor/IDISMessageConverter.java
#	src/main/java/com/energyict/protocolimplv2/messages/convertor/PrimeMeterMessageConverter.java
#	src/main/java/com/energyict/protocolimplv2/messages/convertor/SDKSmartMeterProtocolMessageConverter.java
#	src/main/java/com/energyict/protocolimplv2/messages/convertor/SmartWebRtuKpMessageConverter.java
#	src/main/java/com/energyict/protocolimplv2/messages/convertor/UkHubMessageConverter.java
#	src/main/java/com/energyict/protocolimplv2/messages/convertor/ZigbeeGasMessageConverter.java
#	src/main/java/com/energyict/protocolimplv2/nta/dsmr23/messages/Dsmr23MessageExecutor.java
#	src/main/java/com/energyict/protocolimplv2/nta/dsmr23/messages/Dsmr23Messaging.java
#	src/main/java/com/energyict/protocolimplv2/nta/dsmr40/messages/Dsmr40MessageExecutor.java
#	src/main/java/com/energyict/protocols/messaging/FirmwareUpdateMessageBuilder.java
#	src/main/java/com/energyict/smartmeterprotocolimpl/eict/ukhub/messaging/UkHubMessageExecutor.java
#	src/main/java/com/energyict/smartmeterprotocolimpl/eict/ukhub/zigbee/gas/messaging/ZigbeeMessageExecutor.java
#	src/main/java/com/energyict/smartmeterprotocolimpl/elster/apollo/messaging/AS300MessageExecutor.java
#	src/main/java/com/energyict/smartmeterprotocolimpl/nta/dsmr23/messages/Dsmr23MessageExecutor.java
#	src/main/java/com/energyict/smartmeterprotocolimpl/prenta/iskra/mx372/messaging/IskraMx372Messaging.java

+++ comu/com.energyict.mdc.protocols/src/main/java/com/energyict/protocolimplv2/elster/ctr/MTU155/tariff/rawobjects/RawDayTypeBand.java
@@ -18,11 +18,11 @@ public class RawDayTypeBand extends AbstractField<RawDayTypeBand> {
 
     private static final int LENGTH = 5;
 
-    private int[] code;
+    private long[] code;
     private int[] time;
 
     public RawDayTypeBand() {
-        code = new int[LENGTH];
+        code = new long[LENGTH];
         time = new int[LENGTH];
     }
